requirements.txt
crewai
crewai-tools
pydub
requests>=2.31.0
pydantic
streamlit 
openai
soundfile
ffmpeg
pysqlite3-binary
cloudinary>=1.36.0
python-dotenv>=1.0.0
flask>=3.0.0
